miscellaneous cleanup, pre possible featImplications fix
There may be an error that has slipped by thus far but could be problematic in the future, in that feature implications that are on the mere specification of a feature (as + or -) rather than its specific value may not be functionally computed as implications downstream. At least this seems to be the case in FeatMatrix.applyAlphaValues() -- investigation into whether this is the case in other relevant classes and scripts will commence.